+++ .gitignore
@@ -8,3 +8,4 @@
 /bootstrap/*
 /contacts/*
 /quotes/*
+/test/*
